refactor: router-lite
