+++ packages/router-lite/src/resources/viewport.ts
@@ -1,7 +1,15 @@
 import { ILogger } from '@aurelia/kernel';
-import { LifecycleFlags, bindable, customElement, ICustomElementViewModel, IHydratedController, ICustomElementController, ICompiledCustomElementController } from '@aurelia/runtime-html';
+import {
+  type LifecycleFlags,
+  bindable,
+  customElement,
+  type ICustomElementViewModel,
+  type IHydratedController,
+  type ICustomElementController,
+  type ICompiledCustomElementController
+} from '@aurelia/runtime-html';
 
-import { ViewportAgent } from '../viewport-agent';
+import type { ViewportAgent } from '../viewport-agent';
 import { IRouteContext } from '../route-context';
 import { defaultViewportName } from '../route-definition';
