+++ packages/router-lite/src/viewport-agent.ts
@@ -1,14 +1,15 @@
 // No-fallthrough disabled due to large numbers of false positives
 /* eslint-disable no-fallthrough */
 import { ILogger, onResolve, resolveAll } from '@aurelia/kernel';
-import { LifecycleFlags, IHydratedController, ICustomElementController, Controller } from '@aurelia/runtime-html';
+import { LifecycleFlags, type IHydratedController, type ICustomElementController, Controller } from '@aurelia/runtime-html';
 
-import { IViewport } from './resources/viewport';
-import { ComponentAgent } from './component-agent';
-import { RouteNode, createAndAppendNodes } from './route-tree';
-import { IRouteContext } from './route-context';
-import { Transition, NavigationOptions } from './router';
-import { TransitionPlan } from './route';
+import type { IViewport } from './resources/viewport';
+import type { ComponentAgent } from './component-agent';
+import { type RouteNode, createAndAppendNodes } from './route-tree';
+import type { IRouteContext } from './route-context';
+import type { NavigationOptions } from './options';
+import type { Transition } from './router';
+import type { TransitionPlan } from './route';
 import { Batch, mergeDistinct } from './util';
 import { defaultViewportName } from './route-definition';
 import { ViewportInstruction } from './instructions';
